pyproject.toml
[project]
name = "templatepwa"
version = "0.1.0"
description = "Minimal PWA template with React, aiohttp, and SQLite"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "argon2-cffi>=25.1.0",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.1.1",
    "yoyo-migrations>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-aiohttp>=1.1.0",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 160
target-version = "py314"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
